Include nco-yaml representations of datasets (and script)
As an alternate approach to using nccsv source of truth datasets
in #51, instead we store text yaml representations of the
gold standard datasets based on nco-json output (nco-yaml if you will,
which is more readable than json due to lack of brackets).

Also included is a script to easily (re-)generate the yaml
representations from the source of truth NetCDF files, and
to check if the yaml representations are up to date
(possibly used in a future pre-commit hook, GitHub action, etc).